+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Edit `SentEnvelopeReport` resource
 - Add `AccountBilling.accountContactLimit` property
 - Add `AccountBilling.accountContactUsage` property
+- Add `ContactImportsEndpoint` and `ContactImportItemsEndpoint`
 
 ## [2.9.0] - 2025-10-03
 - [BREAKING-CHANGE] Fix `WebhookAttempt.status` type from string to int